ref(core): Clean up `@sentry/bundler-plugin-core` package (#716)
