+++ packages/bundler-plugin-core/src/logger.ts
