+++ packages/rollup-plugin/test/public-api.test.ts
@@ -22,13 +22,15 @@ describe("sentryRollupPlugin", () => {
 
     const pluginNames = plugins.map((plugin) => plugin.name);
 
-    expect(pluginNames).toEqual([
-      "sentry-telemetry-plugin",
-      "sentry-rollup-release-injection-plugin",
-      "sentry-release-management-plugin",
-      "sentry-rollup-debug-id-injection-plugin",
-      "sentry-rollup-debug-id-upload-plugin",
-      "sentry-file-deletion-plugin",
-    ]);
+    expect(pluginNames).toEqual(
+      expect.arrayContaining([
+        "sentry-telemetry-plugin",
+        "sentry-rollup-release-injection-plugin",
+        "sentry-release-management-plugin",
+        "sentry-rollup-debug-id-injection-plugin",
+        "sentry-rollup-debug-id-upload-plugin",
+        "sentry-file-deletion-plugin",
+      ])
+    );
   });
 });
